Regained some speed from ROI fix.  All out of ROI voxel handling now on GPU.

+++ bspline.h
@@ -84,7 +84,7 @@ struct dev_pointers_bspline
     float* LUT_Bspline_x;
     float* LUT_Bspline_y;
     float* LUT_Bspline_z;
-    int* skipped;			// # of voxels that fell outside post warp
+    float* skipped;			// # of voxels that fell outside post warp
 
     // These hold the size of the
     // chucks of memory we allocated
